商品列表

+++ yudao-module-mall/yudao-module-product-biz/src/main/java/cn/iocoder/yudao/module/product/service/category/ProductCategoryServiceImpl.java
@@ -1,11 +1,14 @@
 package cn.iocoder.yudao.module.product.service.category;
 
+import cn.iocoder.yudao.framework.common.enums.CommonStatusEnum;
 import cn.iocoder.yudao.module.product.controller.admin.category.vo.ProductCategoryCreateReqVO;
 import cn.iocoder.yudao.module.product.controller.admin.category.vo.ProductCategoryListReqVO;
 import cn.iocoder.yudao.module.product.controller.admin.category.vo.ProductCategoryUpdateReqVO;
 import cn.iocoder.yudao.module.product.convert.category.ProductCategoryConvert;
 import cn.iocoder.yudao.module.product.dal.dataobject.category.ProductCategoryDO;
 import cn.iocoder.yudao.module.product.dal.mysql.category.ProductCategoryMapper;
+import cn.iocoder.yudao.module.product.service.spu.ProductSpuService;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
@@ -28,6 +31,9 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
 
     @Resource
     private ProductCategoryMapper productCategoryMapper;
+    @Resource
+    @Lazy // 循环依赖，避免报错
+    private ProductSpuService productSpuService;
 
     @Override
     public Long createCategory(ProductCategoryCreateReqVO createReqVO) {
@@ -61,11 +67,11 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
         if (productCategoryMapper.selectCountByParentId(id) > 0) {
             throw exception(CATEGORY_EXISTS_CHILDREN);
         }
-        // 校验分类是否绑定了 SPU // TODO: 20/11/2023
-//        Long spuCount = productSpuService.getSpuCountByCategoryId(id);
-//        if (spuCount > 0) {
-//            throw exception(CATEGORY_HAVE_BIND_SPU);
-//        }
+        // 校验分类是否绑定了 SPU
+        Long spuCount = productSpuService.getSpuCountByCategoryId(id);
+        if (spuCount > 0) {
+            throw exception(CATEGORY_HAVE_BIND_SPU);
+        }
         // 删除
         productCategoryMapper.deleteById(id);
     }
@@ -96,6 +102,38 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
     @Override
     public ProductCategoryDO getCategory(Long id) {
         return productCategoryMapper.selectById(id);
+    }
+
+    @Override
+    public void validateCategory(Long id) {
+        ProductCategoryDO category = productCategoryMapper.selectById(id);
+        if (category == null) {
+            throw exception(CATEGORY_NOT_EXISTS);
+        }
+        if (Objects.equals(category.getStatus(), CommonStatusEnum.DISABLE.getStatus())) {
+            throw exception(CATEGORY_DISABLED, category.getName());
+        }
+    }
+
+    @Override
+    public Integer getCategoryLevel(Long id) {
+        if (Objects.equals(id, PARENT_ID_NULL)) {
+            return 0;
+        }
+        int level = 1;
+        // for 的原因，是因为避免脏数据，导致可能的死循环。一般不会超过 100 层哈
+        for (int i = 0; i < Byte.MAX_VALUE; i++) {
+            // 如果没有父节点，break 结束
+            ProductCategoryDO category = productCategoryMapper.selectById(id);
+            if (category == null
+                    || Objects.equals(category.getParentId(), PARENT_ID_NULL)) {
+                break;
+            }
+            // 继续递归父节点
+            level++;
+            id = category.getParentId();
+        }
+        return level;
     }
 
     @Override
